Extract features for paw time series

Extract features for paw time series

extract_features(
  x,
  y = NULL,
  parameters = default_parameters(),
  diagnostics = FALSE
)

Arguments

x

time series of horizontal paw movement. Alternatively, a two column matrix of x and y time series, respectively.

y

time series of vertical paw movement, or NULL if x is a two column matrix.

parameters

contains information about frames per second, filtering, windowing, and thresholds, for paw features (see default_parameters, or use set_parameters to modify the defaults).

diagnostics

set to TRUE will record intermediate values used when computing paw features. This information can be helpful for debugging parameter choices. The default, FALSE, is to not record these values.

Value

pre-peak and post-peak paw features (plus diagnostics, if enabled)