Batch convert SLEAP H5 files into DeepLabCut-like CSVs

This function loads a directory containing H5 files generated from SLEAP. It processes the first track in the CSV (assuming single-animal) and rearranges the tracked points to match the style of that in DeepLabCut. Generated CSVs are all saved to a separate directory, but can be saved in the H5 directory if no save_folder parameter is assigned.

h5_to_csv(h5_folder, save_folder = NA)

Arguments

h5_folder

Path to the folder containing SLEAP H5s.

save_folder

Path to the folder where DeepLabCut-like CSVs will be saved.

Value

Converted CSVs stored in the save_folder directory.